+++ CHANGELOG.md
@@ -12,11 +12,12 @@ Anticipated release: July 9, 2021
 
 #### ⚙️ Behind the scenes
 
-
+- Created endpoint to allow a user to switch states ([#3030])
 
 # Previous releases
 
 See our [release history](https://github.com/CMSgov/eAPD/releases)
 
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996
+[#3030]: https://github.com/CMSgov/eAPD/issues/3030
 [#3181]: https://github.com/CMSgov/eAPD/issues/3181